refactor custom commands, add USE custom command
